benchmark/requirements.txt
python-dotenv==1.0.1
pydantic==2.7.3
openai==1.33.0
pandas==2.2.2
xmlschema==3.3.1
tiktoken==0.7.0
